+++ Quartermaster/Assets/Prefabs/World.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 653a5373ac5314140af1090c8895cab5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
