+++ Quartermaster/Assets/Scripts/UIScripts/HPBarScript.cs.meta
